al_demo_execution_client_electron/flask-backend/requirements.txt
flask==3.0.3
websocket-client==1.8.0
selenium==4.25.0
python-dotenv==1.0.1
webdriver-manager==4.0.2
robotframework==7.0.1
robotframework-seleniumlibrary==6.7.1